chore: rename utils-tests folder
