+++ frontend/src/components/common/layouts/__tests__/Container.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { create } from 'react-test-renderer';
 import Container from 'components/common/layouts/Container';
-import withContexts from 'utils-test/withContexts';
+import withContexts from 'utils-tests/withContexts';
 
 describe('Container', () => {
   describe('Given no children', () => {
