+++ frontend/src/components/common/typography/__tests__/AppTitle.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { create } from 'react-test-renderer';
 import AppTitle from 'components/common/typography/AppTitle';
-import withContexts from 'utils-test/withContexts';
+import withContexts from 'utils-tests/withContexts';
 
 describe('AppTitle', () => {
   describe('Given no text', () => {
