+++ frontend/src/pages/__tests__/Home.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { create } from 'react-test-renderer';
 import Home from 'pages/index.page';
-import withContexts from 'utils-test/withContexts';
+import withContexts from 'utils-tests/withContexts';
 
 describe('Home', () => {
   describe('When rendering', () => {
